% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_process_subfuncs.R
\name{check_code_dist}
\alias{check_code_dist}
\title{Base SCV function}
\usage{
check_code_dist(
  cohort,
  concept_set,
  code_type,
  code_domain,
  time = FALSE,
  domain_tbl = read_codeset("scv_domains", "cccc")
)
}
\arguments{
\item{cohort}{table of cohort members with at least \code{site}, \code{person_id}, \code{start_date}, and \code{end_date}}

\item{concept_set}{for analyses where time = FALSE, a csv file with the source or cdm codes of interest for the analysis.
should contain at least a \code{concept_id} column

\if{html}{\out{<div class="sourceCode">}}\preformatted{               for analyses where time = TRUE, a vector with up to 5 source or cdm codes of interest for the analysis.
}\if{html}{\out{</div>}}}

\item{code_type}{the type of code to be examined in the check; either \code{source} or \code{cdm}}

\item{code_domain}{the domain related to the codes in the \code{concept_set}; should match a domain and
its associated metadata in the \code{domain_tbl} file}

\item{time}{logical to indicate whether the user would like to examine mappings over time or not}

\item{domain_tbl}{a table with a list of domains and associated metadata; should include the name
of the domain, the name of the column where source codes can be found, the name
of the column where CDM codes can be found, and the date column that should
be used during over time computations}
}
\value{
one dataframe with counts and proportions of source -> cdm or cdm -> source mapping pairs
for each facet group. if time = FALSE, these counts are computed overall. if time = TRUE,
these counts are computed for each user-specified time increment.
}
\description{
Base SCV function
}
