% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_process_subfuncs.R
\name{compute_jaccard_scv}
\alias{compute_jaccard_scv}
\title{Compute Jaccard at the visit level}
\usage{
compute_jaccard_scv(
  cohort,
  domain_tbl,
  concept_set,
  code_type,
  code_domain,
  omop_or_pcornet
)
}
\arguments{
\item{cohort}{table of cohort members with at least \code{site},
a patient identifier, \code{start_date}, and \code{end_date}}

\item{domain_tbl}{a table with a list of domains and associated metadata; should include the name
of the domain, the name of the column where source codes can be found, the name
of the column where CDM codes can be found, and the date column that should
be used during over time computations}

\item{concept_set}{a csv file with the source or cdm codes of interest for the analysis.
should contain at least a \code{concept_id} column=}

\item{code_type}{the type of code to be examined in the check; either \code{source} or \code{cdm}}

\item{code_domain}{the domain related to the codes in the \code{concept_set}; should match a domain and
its associated metadata in the \code{domain_tbl} file}

\item{omop_or_pcornet}{Option to run the function using the OMOP or PCORnet CDM as the default CDM}
}
\value{
a dataframe summarizing, for each code in the concept set, the jaccard
similarity index of two mapped concepts within the same visit
}
\description{
This function will compute the Jaccard similarity index
at the visit level to determine how often two mappings occur
within the same visit when mapped to/from a given concept
}
\keyword{internal}
