% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_output_subfuncs.R
\name{compute_mappings_per_code}
\alias{compute_mappings_per_code}
\title{Compute number and median mappings per code}
\usage{
compute_mappings_per_code(tbl, col, denom, facet = NULL)
}
\arguments{
\item{tbl}{intermediate table generated in the output function that contains the concepts
of interest to be used to compute number of mappings}

\item{col}{the name of the column with the concept that needs to be summarized}

\item{denom}{the denominator count associated with @col}

\item{facet}{grouping variables to be used to compute group-specific counts and distance
from overall median}
}
\value{
dataframe that summarizes the overall median number of mappings, group-specific number of
mappings, and how many MAD from the overall median a code falls
}
\description{
Compute number and median mappings per code
}
