% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scv_domain_file}
\alias{scv_domain_file}
\title{SCV Sample Domain File}
\format{
\subsection{\code{scv_domain_file}}{

A data frame with 2 rows and 5 columns
\describe{
\item{domain}{The name of the CDM table associated with the domain where the concept of interest can be found. Should match the domain listed in the concept_set file.}
\item{concept_field}{The name of the column in the domain table that contains the concepts of interest}
\item{source_field}{The name of the column in the domain table that contains the source concepts of interest}
\item{date_field}{The name of the column in the domain table that contains dates to be used for time-based filtering.}
\item{vocabulary_field}{(PCORnet only) The name of the column in the domain table where the vocabulary type is stored}
}
}
}
\usage{
scv_domain_file
}
\description{
A sample version of the file structure expected for the \code{domain_tbl}
parameter in the \code{scv_process} function. The user should recreate this file
and include their own domain definitions.
}
\keyword{datasets}
