% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_output_subfuncs.R
\name{scv_ms_anom_at}
\alias{scv_ms_anom_at}
\title{\strong{Multi-Site Across Time Anomaly}
Produces graphs showing Euclidean Distanctes}
\usage{
scv_ms_anom_at(process_output, code_type, filter_concept, filter_mapped)
}
\arguments{
\item{process_output}{output from \code{evp_process}}

\item{code_type}{type of code to be used in analysis -- either \code{source} or \code{cdm}

\if{html}{\out{<div class="sourceCode">}}\preformatted{             should match the code_type provided when running `scv_process`
}\if{html}{\out{</div>}}}

\item{filter_concept}{the code_type concept of interest to be displayed in the output}

\item{filter_mapped}{the mapped concept of interest to be displayed in the output}
}
\value{
three graphs:
\enumerate{
\item line graph that shows the smoothed proportion of a
concept pair across time computation with the Euclidean distance associated with each line
\item line graph that shows the raw proportion of a
concept pair across time computation with the Euclidean distance associated with each line
\item a bar graph with the Euclidean distance value for each site, with the average
proportion as the fill
}

THIS GRAPH SHOWS ONLY ONE VARIABLE AT A TIME!
}
\description{
\strong{Multi-Site Across Time Anomaly}
Produces graphs showing Euclidean Distanctes
}
