% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_output_subfuncs.R
\name{scv_ms_anom_cs}
\alias{scv_ms_anom_cs}
\title{\emph{Multi-Site, Anomaly, Cross-Sectional}}
\usage{
scv_ms_anom_cs(
  process_output,
  code_type,
  facet = NULL,
  filter_concept,
  num_mappings = 20
)
}
\arguments{
\item{process_output}{dataframe output by \code{scv_process}}

\item{code_type}{type of code to be used in analysis -- either \code{source} or \code{cdm};
should match the code_type provided when running \code{scv_process}}

\item{facet}{the variables by which you would like to facet the graph}

\item{filter_concept}{the code_type concept_id of interest for which to display its mappings}

\item{num_mappings}{an integer indicating the number of top mappings that should be displayed}
}
\value{
a dot plot where the shape of the dot represents whether the point is
anomalous, the color of the dot represents the proportion of representation
of a filter_concept/mapping concept pair, and the size of the dot
represents the mean proportion across all sites
}
\description{
\emph{Multi-Site, Anomaly, Cross-Sectional}
}
