% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_output_subfuncs.R
\name{scv_ms_anom_la}
\alias{scv_ms_anom_la}
\title{\strong{Multi-Site, Anomaly, Longitudinal}}
\usage{
scv_ms_anom_la(
  process_output,
  code_type,
  filter_concept,
  filter_mapped,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{output from \code{evp_process}}

\item{code_type}{type of code to be used in analysis -- either \code{source} or \code{cdm};
should match the code_type provided when running \code{scv_process}}

\item{filter_concept}{the code_type concept of interest to be displayed in the output}

\item{filter_mapped}{the mapped concept of interest to be displayed in the output}

\item{large_n}{a boolean indicating whether the large N visualization, intended for a high
volume of sites, should be used; defaults to FALSE}

\item{large_n_sites}{a vector of site names that can optionally generate a filtered visualization}
}
\value{
three graphs:
\enumerate{
\item line graph that shows the smoothed proportion of a
concept pair across time computation with the Euclidean distance associated with each line
\item line graph that shows the raw proportion of a
concept pair across time computation with the Euclidean distance associated with each line
\item a bar graph with the Euclidean distance value for each site, with the average
proportion as the fill
}

THIS GRAPH SHOWS ONLY ONE CONCEPT PAIR AT A TIME!
}
\description{
\strong{Multi-Site, Anomaly, Longitudinal}
}
\keyword{internal}
