% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_output_subfuncs.R
\name{scv_ms_exp_la}
\alias{scv_ms_exp_la}
\title{\emph{Multi Site, Exploratory, Longitudinal}}
\usage{
scv_ms_exp_la(
  process_output,
  code_type,
  filter_concept,
  num_mappings = 10,
  facet = NULL
)
}
\arguments{
\item{process_output}{dataframe output by \code{scv_process}}

\item{code_type}{type of code to be used in analysis -- either \code{source} or \code{cdm};
should match the code_type provided when running \code{scv_process}}

\item{filter_concept}{the code_type concept of interest for which mappings should be shown}

\item{num_mappings}{an integer indicating the top number of mappings for filter_concept
that should be displayed}

\item{facet}{the variables by which you would like to facet the graph}
}
\value{
a line graph with one facet per code displaying the proportion of mapped codes
across the user selected time period

a reference table with total counts of each code across the entire user selected
time period
}
\description{
\emph{Multi Site, Exploratory, Longitudinal}
}
