% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_output_subfuncs.R
\name{scv_ms_exp_nt}
\alias{scv_ms_exp_nt}
\title{\emph{Multi Site, Exploratory, No Time}}
\usage{
scv_ms_exp_nt(process_output, code_type, facet = NULL, num_codes = 10)
}
\arguments{
\item{process_output}{dataframe output by \code{scv_process}}

\item{code_type}{type of code to be used in analysis -- either \code{source} or \code{cdm}

\if{html}{\out{<div class="sourceCode">}}\preformatted{             should match the code_type provided when running `scv_process`
}\if{html}{\out{</div>}}}

\item{facet}{the variables by which you would like to facet the graph}

\item{num_codes}{the number of top codes of code_type that should be displayed in the graph}
}
\value{
a searchable and filterable table with mappings, proportion of representation, and
denominator counts for the number of codes selected
in @num_codes
concept name will be included if @vocab_tbl is not NULL
}
\description{
\emph{Multi Site, Exploratory, No Time}
}
