% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_output.R
\name{scv_output}
\alias{scv_output}
\title{Source and Concept Vocabularies -- Output Generation}
\usage{
scv_output(
  process_output,
  code_type,
  filter_concept = NULL,
  filter_mapped = NULL,
  num_codes = 10,
  num_mappings = 25,
  vocab_tbl = NULL,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{\emph{tabular input} || \strong{required}

The tabular output produced by \code{scv_process}}

\item{code_type}{\emph{string} || \strong{required}

A string identifying the type of concept that was been provided in the \code{concept_set}. This
should match whatever was provided when executing the \code{scv_process} function.

Acceptable values are \code{cdm} (the standard, mapped code that is included in the CDM) or \code{source} (the "raw" concept from the source system)}

\item{filter_concept}{\emph{(integer or string) or vector} || defaults to \code{NULL}

An integer/string (depending on the data model) or vector of concepts (as provided in the \code{concept_set} in \code{scv_process})
that should appear in the output. This parameter is required for the following check types:
\itemize{
\item \verb{Single Site, Anomaly Detection, Cross-Sectional}
\item \verb{Multi-Site, Anomaly Detection, Cross-Sectional}
\item \verb{Multi-Site, Exploratory, Longitudinal}
\item \verb{Single Site, Anomaly Detection, Longitudinal}
\item \verb{Multi-Site, Anomaly Detection, Longitudinal}
}}

\item{filter_mapped}{\emph{integer or string} || defaults to \code{NULL}

An integer/string (depending on the data model) representing a mapping associated with \code{filter_concept},
as it appears in the output. This parameter is required for the following check types:
\itemize{
\item \verb{Multi-Site, Anomaly Detection, Longitudinal}
}}

\item{num_codes}{\emph{integer} || defaults to \code{10}

An integer indicating the top N of concepts (of \code{code_type}) that should be
displayed in the plot. This parameter is required for the following check types:
\itemize{
\item \verb{Single Site, Exploratory, Cross-Sectional}
\item \verb{Multi-Site, Exploratory, Cross-Sectional}
\item \verb{Single Site, Anomaly Detection, Cross-Sectional}
}}

\item{num_mappings}{\emph{integer} || defaults to \code{25}

An integer indicating the top N of mappings associated with each concept of \code{code_type}
that should be displayed on the plot. This parameter is required for the following
check types:
\itemize{
\item \verb{Single Site, Exploratory, Cross-Sectional}
\item \verb{Single Site, Anomaly Detection, Cross-Sectional}
\item \verb{Multi-Site, Anomaly Detection, Cross-Sectional}
\item \verb{Single Site, Exploratory, Longitudinal}
\item \verb{Multi-Site, Exploratory, Longitudinal}
\item \verb{Multi-Site, Anomaly Detection, Longitudinal}
}}

\item{vocab_tbl}{\emph{tabular input} || \strong{optional}

A vocabulary table containing concept names for the provided codes
(ex: the OMOP concept table)}

\item{large_n}{\emph{boolean} || defaults to \code{FALSE}

For Multi-Site analyses, a boolean indicating whether the large N
visualization, intended for a high volume of sites, should be used. This
visualization will produce high level summaries across all sites, with an
option to add specific site comparators via the \code{large_n_sites} parameter.}

\item{large_n_sites}{\emph{vector} || defaults to \code{NULL}

When \code{large_n = TRUE}, a vector of site names that can add site-level information
to the plot for comparison across the high level summary information.}
}
\value{
This function will produce a graph to visualize the results
from \code{scv_process} based on the parameters provided. The default
output is typically a static ggplot or gt object, but interactive
elements can be activated by passing the plot through \code{make_interactive_squba}.
For a more detailed description of output specific to each check type,
see the PEDSpace metadata repository
}
\description{
Using the tabular output generated by \code{scv_process}, this function will build a graph to
visualize the results. Each function configuration will output a bespoke ggplot. Theming can
be adjusted by the user after the graph has been output using \code{+ theme()}. Most graphs can
also be made interactive using \code{make_interactive_squba()}
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'sourceconceptvocabularies'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'scv_process_test',
                      working_directory = my_directory,
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = my_file_folder,
                      cdm_schema = NA)

#' Build mock study cohort
cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000),
                #RSQLite does not store date objects,
                #hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Prepare input tables
scv_domain_tbl <- dplyr::tibble(domain = 'condition_occurrence',
                                concept_field = 'condition_concept_id',
                                source_concept_field =
                                  'condition_source_concept_id',
                                date_field = 'condition_start_date',
                                vocabulary_field = NA)

scv_concept_set <- read_codeset('dx_hypertension')

#' Execute `scv_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
scv_process_example <- scv_process(cohort = cohort,
                                   multi_or_single_site = 'single',
                                   anomaly_or_exploratory = 'exploratory',
                                   time = FALSE,
                                   omop_or_pcornet = 'omop',
                                   code_type = 'cdm',
                                   code_domain = 'condition_occurrence',
                                   domain_tbl = scv_domain_tbl,
                                   concept_set = scv_concept_set) \%>\%
  suppressMessages()

scv_process_example

#' Execute `scv_output` function
scv_output_example <- scv_output(process_output = scv_process_example,
                                 code_type = 'cdm',
                                 vocab_tbl = NULL) \%>\%
  suppressMessages()

scv_output_example[[1]]

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(scv_output_example[[1]])
}
