% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_output_subfuncs.R
\name{scv_ss_anom_cs}
\alias{scv_ss_anom_cs}
\title{\emph{Single Site, Anomaly, Cross-Sectional}}
\usage{
scv_ss_anom_cs(
  process_output,
  code_type,
  facet = NULL,
  num_codes = 10,
  num_mappings = 15
)
}
\arguments{
\item{process_output}{dataframe output by \code{scv_process}}

\item{code_type}{type of code to be used in analysis -- either \code{source} or \code{cdm};
should match the code_type provided when running \code{scv_process}}

\item{facet}{the variables by which you would like to facet the graph}

\item{num_codes}{the number of top codes of code_type that should be displayed in the graph}

\item{num_mappings}{the number of top mappings that should be displayed for each code}
}
\value{
a dot plot where the shape of the dot represents whether the point is
anomalous, the color of the dot represents the proportion of usage
for a cdm/source concept pair, and the size of the dot represents the
mean proportion for the code type along the x axis
}
\description{
\emph{Single Site, Anomaly, Cross-Sectional}
}
