% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_output_subfuncs.R
\name{scv_ss_anom_la}
\alias{scv_ss_anom_la}
\title{\emph{Single Site, Anomaly, Longitudinal}}
\usage{
scv_ss_anom_la(process_output, code_type, filter_concept, facet = NULL)
}
\arguments{
\item{process_output}{dataframe output by \code{scv_process}}

\item{code_type}{type of code to be used in analysis -- either \code{source} or \code{cdm};
should match the code_type provided when running \code{scv_process}}

\item{filter_concept}{the code_type concept of interest to be displayed in the output}

\item{facet}{the variables by which you would like to facet the graph}
}
\value{
if analysis was executed by year or greater, a C control chart
showing the number of mappings for filter_concept
is returned with outliers marked with orange dots

\if{html}{\out{<div class="sourceCode">}}\preformatted{    if analysis was executed by month or smaller, an STL regression is
    conducted and outliers are marked with red dots. the graphs representing
    the data removed in the regression are also returned
}\if{html}{\out{</div>}}
}
\description{
\emph{Single Site, Anomaly, Longitudinal}
}
\keyword{internal}
