% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_output_subfuncs.R
\name{scv_ss_exp_la}
\alias{scv_ss_exp_la}
\title{\emph{Single Site, Exploratory, Longitudinal}}
\usage{
scv_ss_exp_la(process_output, code_type, num_mappings = 10, facet = NULL)
}
\arguments{
\item{process_output}{dataframe output by \code{scv_process}}

\item{code_type}{type of code to be used in analysis -- either \code{source} or \code{cdm};
should match the code_type provided when running \code{scv_process}}

\item{num_mappings}{integer indicating the number of top mappings per code to be
displayed in the output}

\item{facet}{the variables by which you would like to facet the graph}
}
\value{
a line graph with one facet per code displaying the proportion of mapped codes
across the user selected time period

a reference table with total counts of each code across the entire user selected
time period
}
\description{
\emph{Single Site, Exploratory, Longitudinal}
}
