% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_output_subfuncs.R
\name{scv_ss_exp_nt}
\alias{scv_ss_exp_nt}
\title{\emph{Single Site, Exploratory, No Time}}
\usage{
scv_ss_exp_nt(
  process_output,
  code_type,
  facet = NULL,
  num_codes = 10,
  num_mappings = 10
)
}
\arguments{
\item{process_output}{dataframe output by \code{scv_process}}

\item{code_type}{type of code to be used in analysis -- either \code{source} or \code{cdm}

\if{html}{\out{<div class="sourceCode">}}\preformatted{             should match the code_type provided when running `scv_process`
}\if{html}{\out{</div>}}}

\item{facet}{the variables by which you would like to facet the graph}

\item{num_codes}{the number of top codes of code_type that should be displayed in the graph}

\item{num_mappings}{the number of top mappings that should be displayed for each code}
}
\value{
a heatmap with one facet per code and additional facet groupings, limited to the number
of codes selected with @num_codes
mapped code along the y-axis and proportion of representation as the fill

a reference table with additional information about the codes used as the facet.
includes the code and denominator count, and will also include the concept name
if @vocab_tbl is not NULL
}
\description{
\emph{Single Site, Exploratory, No Time}
}
