% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_output.R
\name{scv_output}
\alias{scv_output}
\title{SCV Output Generation}
\usage{
scv_output(
  process_output,
  output_function,
  code_type,
  facet = NULL,
  filter_concept = NULL,
  filter_mapped = NULL,
  num_codes = 10,
  num_mappings = 25,
  vocab_tbl = vocabulary_tbl("concept")
)
}
\arguments{
\item{process_output}{the output of the \code{scv_process} function}

\item{output_function}{the name of the output function that should be used provided in the \code{parameter_summary} csv
file that is output to the provided results folder after running the \code{scv_process} function}

\item{code_type}{the type of code that is being used in the analysis, either \code{source} or \code{cdm}

\if{html}{\out{<div class="sourceCode">}}\preformatted{             should ideally match the code_type that was defined when running `scv_process`
}\if{html}{\out{</div>}}}

\item{facet}{the variables by which you would like to facet the graph. available and/or recommended options for
faceting variables are provided in the \code{parameter_summary} csv file}

\item{filter_concept}{for \code{scv_ms_anom_at} only -- choose ONE concept_id from the concept_set provided in
\code{scv_process} to filter the output}

\item{filter_mapped}{for \code{scv_ms_anom_at} only -- choose ONE mapped concept from those associated with the
concept_id provided in \code{filter_concept}; options can be found in the \code{mapped_id} column
of \code{scv_process}}

\item{num_codes}{the number of top codes of code_type that should be displayed in the analysis

\if{html}{\out{<div class="sourceCode">}}\preformatted{             used for `ss_exp_nt` and `ms_exp_nt`
}\if{html}{\out{</div>}}}

\item{num_mappings}{the number of top mappings that should be displayed for each code of code_type

\if{html}{\out{<div class="sourceCode">}}\preformatted{                used for `ss_exp_nt`
}\if{html}{\out{</div>}}}

\item{vocab_tbl}{OPTIONAL: the location of an external vocabulary table containing concept names for
the provided codes. if not NULL, concept names will be available in either a reference
table or in a hover tooltip}
}
\value{
a graph to visualize the results from \code{scv_process} based on the parameters provided

\if{html}{\out{<div class="sourceCode">}}\preformatted{    in some cases, an additional reference table with summary information about the codes
    included in the graph
}\if{html}{\out{</div>}}
}
\description{
SCV Output Generation
}
