% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_process.R
\name{scv_process}
\alias{scv_process}
\title{SCV Process function}
\usage{
scv_process(
  cohort,
  concept_set,
  domain_tbl = sourceconceptvocabularies::scv_domain_file,
  code_type = "source",
  code_domain = "condition_occurrence",
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  p_value = 0.9,
  age_groups = NULL,
  time = FALSE,
  time_span = c("2012-01-01", "2020-01-01"),
  time_period = "year"
)
}
\arguments{
\item{cohort}{A dataframe with the cohort of patients for your study. Should include the columns:
- @person_id
- @start_date
- @end_date
- @site}

\item{concept_set}{for analyses where time = FALSE, a csv file with the source or cdm codes of interest for the analysis.
should contain at least a \code{concept_id} column

\if{html}{\out{<div class="sourceCode">}}\preformatted{               for analyses where time = TRUE, a vector with up to 5 source or cdm codes of interest for the analysis.
}\if{html}{\out{</div>}}}

\item{domain_tbl}{a csv file that defines the domains where facts should be identified. defaults to the provided
\code{scv_domains.csv} file, which contains the following fields:
- @domain: the CDM table where information for this domain can be found (i.e. drug_exposure)
- @source_col: the column in the CDM table where \code{source} codes can be identified (i.e. drug_source_concept_id)
- @concept_col: the column in the CDM table where \code{cdm} codes can be identified (i.e. drug_concept_id)
- @date_col: the column in the CDM table that should be used as the default date field for
over time analyses (i.e. drug_exposure_start_date)}

\item{code_type}{the type of code that is being used in the analysis, either \code{source} or \code{cdm}}

\item{code_domain}{the domain where the codes in the concept set should be searched for; must match
a domain defined in @domain_tbl}

\item{multi_or_single_site}{Option to run the function on a single vs multiple sites
- @single - run the function for a single site
- @multi - run the function for multiple sites}

\item{anomaly_or_exploratory}{Option to conduct an exploratory or anomaly detection analysis. Exploratory analyses give a high
level summary of the data to examine the fact representation within the cohort. Anomaly detection
analyses are specialized to identify outliers within the cohort.}

\item{p_value}{the p value to be used as a threshold in the multi-site anomaly detection analysis}

\item{age_groups}{If you would like to stratify the results by age group, fill out the provided \code{age_group_definitions.csv} file
with the following information:
- @min_age: the minimum age for the group (i.e. 10)
- @max_age: the maximum age for the group (i.e. 20)
- @group: a string label for the group (i.e. 10-20, Young Adult, etc.)

\if{html}{\out{<div class="sourceCode">}}\preformatted{                Then supply this csv file as the age_groups argument (i.e. read.csv('path/to/age_group_definitions.csv'))

                If you would not like to stratify by age group, leave the argument as NULL
}\if{html}{\out{</div>}}}

\item{time}{a logical that tells the function whether you would like to look at the output over time}

\item{time_span}{when time = TRUE, this argument defines the start and end dates for the time period of interest. should be
formatted as c(start date, end date) in yyyy-mm-dd date format}

\item{time_period}{when time = TRUE, this argument defines the distance between dates within the specified time period. defaults
to \code{year}, but other time periods such as \code{month} or \code{week} are also acceptable}
}
\value{
a dataframe with counts and proportions for each source -> cdm or cdm -> source mapping
pair for each of the codes provided in @concept_set. this output should then be used in
the \code{scv_output} function to generate an appropriate visualization
}
\description{
SCV Process function
}
